fix(github_actions): Skip failed boilerplate commit
This ports the change from #3317.

+++ .github/scripts/apidoc.sh
@@ -100,15 +100,15 @@ git commit -m "[relevant] $COMMIT_MSG"
 mv .git ../doc
 cd ../doc
 git add --all .
-git commit -m "[boilerplate] $COMMIT_MSG"
-
-# Reorder/swap commits, to have "relevant" after "boilerplate".
-# This makes it show up earlier in the Github interface etc.
-git tag _old
-git reset --hard HEAD~2
-git cherry-pick _old _old~1
-RELEVANT_REV="$(git rev-parse --short HEAD)"
-git tag -d _old
+if git commit -m "[boilerplate] $COMMIT_MSG"; then
+    # Reorder/swap commits, to have "relevant" after "boilerplate".
+    # This makes it show up earlier in the Github interface etc.
+    git tag _old
+    git reset --hard HEAD~2
+    git cherry-pick _old _old~1
+    RELEVANT_REV="$(git rev-parse --short HEAD)"
+    git tag -d _old
+fi
 
 git checkout "$BRANCH"
 OLD_REV="$(git rev-parse --short HEAD)"
